+++ src/AbstractPsrLogger.php
@@ -106,4 +106,18 @@ abstract class AbstractPsrLogger implements LoggerInterface
 
         $this->getYiiLogger()->log($errorMessage, CLogger::LEVEL_ERROR, 'system.log');
     }
+
+    /**
+     * Creates new self instance.
+     * This method can be useful when writing chain methods calls.
+     *
+     * @since 1.0.1
+     *
+     * @param mixed ...$args constructor arguments.
+     * @return static new self instance.
+     */
+    public static function new(...$args): self
+    {
+        return new static(...$args);
+    }
 }