+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Yii1 PSR Log extension
 -----------------------
 
 - Enh: Added ability for global log context setup at `PsrLogger` (klimov-paul)
+- Enh: Added `PsrLogger::new()` static method for a new instance creation (klimov-paul)
 
 
 1.0.0, July 19, 2023